% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPKG.R
\name{NPKG}
\alias{NPKG}
\title{NPKG (Network - Phenotype Integrated Key Gene screening)}
\usage{
NPKG(
  gene_data,
  pheno_data,
  pheno,
  pheno_group_fn,
  p_threshold = 0.05,
  gs_threshold = 0.25,
  mm_threshold = 0.75,
  degree_top = 0.05,
  clust_threshold = 0.95,
  output_dir = ".",
  seed = 423,
  verbose = TRUE
)
}
\arguments{
\item{gene_data}{Preprocessed gene expression matrix (genes in columns, samples in rows)}

\item{pheno_data}{Phenotype data frame containing sample information and traits}

\item{pheno}{Phenotype trait name}

\item{pheno_group_fn}{Phenotype grouping function}

\item{p_threshold}{Module-trait correlation p-value threshold (default = 0.05)}

\item{gs_threshold}{Gene Significance threshold (default = 0.25)}

\item{mm_threshold}{Module Membership threshold (default = 0.75)}

\item{degree_top}{Hub genes degree threshold (default = 0.05)}

\item{clust_threshold}{Remove outlier samples through clustering (default = 0.95)}

\item{output_dir}{Output directory}

\item{seed}{Random Seed}

\item{verbose}{Whether to show progress messages}
}
\value{
\itemize{
\item modules - significantly phenotype_related modules
\item final_genes - Key genes in the final screen
}
}
\description{
NPKG (Network - Phenotype Integrated Key Gene screening)
}
